+++ resources/views/components/avatar/index.blade.php
@@ -1,10 +1,13 @@
 <div data-slot="avatar" data-size="{{ $size }}" {{ $attributes->class([
-    'block relative rounded-full overflow-hidden aspect-square',
+    'block relative overflow-hidden aspect-square',
     'bg-[var(--secondary)] text-[var(--muted-foreground)]',
     'ring-1 ring-[var(--background)]',
 
     // Border element, will be overlain with the image.
-    'before:content-[\'\'] before:absolute before:inset-0 before:rounded-full before:border before:border-[var(--border)]',
+    'before:content-[\'\'] before:absolute before:inset-0 before:border before:border-[var(--border)]',
+
+    'rounded-[var(--radius)] before:rounded-[var(--radius)]' => ! $circle,
+    'rounded-full before:rounded-full' => $circle,
 
     '[&_[data-slot=icon]]:size-4 w-10' => $size === 'lg',
     '[&_[data-slot=icon]]:size-4 w-9' => $size === 'default',